pyproject.toml
[project]
name = "ai-experiments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.2"
dependencies = [
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "ollama>=0.5.1",
    "openai>=1.84.0",
    "pydantic>=2.11.5",
    "pytest>=8.4.0",
    "pytest-cov>=6.1.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "uvicorn[standard] (>=0.34.3,<0.35.0)",
    "fastapi-mcp (>=0.3.4,<0.4.0)",
    "pyperclip>=1.9.0",
    "fitz>=0.0.1.dev2",
    "pymupdf>=1.26.3",
    "langchain>=0.3.27",
    "gradio>=5.39.0",
    "langchain-community>=0.3.27",
    "sentence-transformers>=5.0.0",
    "faiss-cpu>=1.11.0.post1",
]

[tool.uv.workspace]
members = ["scaffolder"]
